Initial refactor of invites to modular system
